♻️ refactor: Melhora validações em DTOs e ValueObjects
- Adiciona validações de tipo em CreateArticleDTO::fromArray()
- Adiciona validações de tipo em CreateUserDTO::fromArray()
- Melhora mensagens de erro em Content, Email, Name e Url
- Garante type safety nas conversões de array para DTO

+++ app/ValueObjects/Content.php
@@ -47,7 +47,7 @@ final readonly class Content implements Stringable
 
     public function getReadingTime(int $wordsPerMinute = 200): int
     {
-        return max(1, ceil($this->getWordCount() / $wordsPerMinute));
+        return max(1, (int) ceil($this->getWordCount() / $wordsPerMinute));
     }
 
     public function isEmpty(): bool
